fix_and_add_some_methods
add getsByDate and getAllByDate in DishController to get dishes on a certain day from one or all restaurants. Add get vote and getAll votes for authorized user in VoteController and fix method vote in VoteService. Get rid of some instant variables in Restaurant and fixed the methods that depended on them